requirements.txt
streamlit==1.23.1
mysql-connector-python==8.1.0
pandas==1.5.3
plotly==5.15.0